[JENKINS-13154] removed hot lock contention
annotatedTypes is used to keep track of classes whose annotations we've already processed. So on a sufficiently warm system, annotatedTypes.contains(type) should almost always return true. Thus we just need to move it out of the synchronization block to be able to eliminate the contention. For this, we use ConcurrentWeakHashMap instead of regular WeakHashMap.

The processTypes method still needs to run in a synchronized block to prevent concurrent calls of processAnnotation with the same class to serialize their executions. It is safe to enter this synchronized block after the initial annotatedTypes.contains(type) check, as the processTypes method itself guards against repeated processing of the same type.